Fix scanner regressions and package ownership
