+++ packages/core/src/sparql-cursors.ts
@@ -1,19 +1,8 @@
-import {
-  decodeSparqlCodePointEscapes,
-  readSparqlVariable,
-  skipSparqlIriRefForStructuralScan,
-  skipSparqlSpaceAndLineComments,
-  skipSparqlStringLiteral,
-} from './sparql-lexical-primitives.js';
-
-/** Decode SPARQL UCHAR escapes before handing a request to a parser. */
-export function preprocessSparqlCodePointEscapes(source: string): string | null {
-  return decodeSparqlCodePointEscapes(source);
-}
-
 export {
+  preprocessSparqlCodePointEscapes,
   readSparqlVariable,
+  skipSparqlIriRef,
   skipSparqlIriRefForStructuralScan,
   skipSparqlSpaceAndLineComments,
   skipSparqlStringLiteral,
-};
+} from '@origintrail-official/dkg-rdf-utils/sparql-cursors';
