+++ packages/rdf-utils/README.md
@@ -1,4 +1,4 @@
 # DKG RDF utilities
 
-Dependency-free RDF serialization helpers shared by DKG packages that must not
-depend on the full DKG core runtime.
+Dependency-free RDF serialization, SPARQL lexical, and bounded-response helpers
+shared by DKG packages that must not depend on the full DKG core runtime.
