+++ packages/rdf-utils/package.json
@@ -10,6 +10,21 @@
       "import": "./dist/index.js",
       "default": "./dist/index.js"
     },
+    "./bounded-response-body": {
+      "types": "./dist/bounded-response-body.d.ts",
+      "import": "./dist/bounded-response-body.js",
+      "default": "./dist/bounded-response-body.js"
+    },
+    "./sparql-lexical-scanner": {
+      "types": "./dist/sparql-lexical-scanner.d.ts",
+      "import": "./dist/sparql-lexical-scanner.js",
+      "default": "./dist/sparql-lexical-scanner.js"
+    },
+    "./sparql-cursors": {
+      "types": "./dist/sparql-cursors.d.ts",
+      "import": "./dist/sparql-cursors.js",
+      "default": "./dist/sparql-cursors.js"
+    },
     "./package.json": "./package.json"
   },
   "scripts": {
